requirements.txt
grpcio
grpcio-tools
regex
websocket-client
